// src/app/Interfaces/View/all-columns-customer-product-group.ts
export interface AllColumnsCustomerProductGroup {
    customerId: number;
    customerName: string;
    territoryName: string;
    sectorName: string;
    governorateName: string;
    regionName: string;
    productGroupName: string;
    productGroupID: number;
    company : string;
    adress : string;
    territoryID: number;
    sectorID: number;
    governorateID: number;
    regionID: number;
    total:number;
}